Moved MakeNewPRing...  to GOperations.C
There is a big block of functions related to converting data from a "GOperation" to the internal call to its "GReductor".
I believe they make more sense inside TmpGOperation.C (and inside an anonymous namespace!).
I have not (yet) updated such code, but only adjusted the order of the functions because of their current lack of forward declarations (inside anon namespace).
First step: MakeNewPRing... (for representing a free module into a ring)
